# api/admin.openapi.yaml
openapi: 3.1.0
info:
  title: Sidervia Admin API
  version: 0.1.0
  description: Same-origin control-plane API for a single Sidervia administrator.
servers:
  - url: /api/admin/v1
security:
  - sessionCookie: []
paths:
...
  /auth/password:
    put:
      operationId: changePassword
      parameters:
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '200':
          description: Rotated current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        default:
          $ref: '#/components/responses/Error'
  /auth/totp/setup:
    post:
      operationId: setupTOTP
      parameters:
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password]
              properties:
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Pending TOTP secret, returned only during setup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TOTPSetup'
        default:
          $ref: '#/components/responses/Error'
  /auth/totp/confirm:
    post:
      operationId: confirmTOTP
      parameters:
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TOTPCodeRequest'
      responses:
        '200':
          description: Rotated current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        default:
          $ref: '#/components/responses/Error'
  /auth/totp:
    delete:
      operationId: disableTOTP
      parameters:
        - $ref: '#/components/parameters/CSRF'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password, code]
              properties:
                password:
                  type: string
                  format: password
                code:
                  type: string
                  pattern: '^[0-9]{6}$'
      responses:
        '200':
          description: Rotated current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sidervia_session
  parameters:
    CSRF:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Stable safe error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
...
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
          minLength: 14
          maxLength: 256
    TOTPCodeRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code:
          type: string
          pattern: '^[0-9]{6}$'
    Session:
      type: object
      additionalProperties: false
      required: [csrf_token, created_at, idle_expires_at, absolute_expires_at, totp_enabled]
      properties:
        csrf_token:
          type: string
        created_at:
          type: string
          format: date-time
        idle_expires_at:
          type: string
          format: date-time
        absolute_expires_at:
          type: string
          format: date-time
        totp_enabled:
          type: boolean
    TOTPSetup:
      type: object
      additionalProperties: false
      required: [secret, uri, expires_at]
      properties:
        secret:
          type: string
        uri:
          type: string
        expires_at:
          type: string
          format: date-time
...